src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "configparser>=7.2.0",
    "fastapi[standard]>=0.115.12",
    "mysqlclient>=2.2.7",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
